initial commit
Create initial structure for rws2wav tool. I haven't figured out the
file format yet (the documentation I do have is for a different version
and apparently isn't correct), so for now this is only adequate to read
the track count. As I'm not expecting this to become a large project,
and don't especially care about portability at the moment, it is
currently using an in-source build system via a trivial Makefile. (If
this grows more complicated, I may port it to CMake.)