read remaining headers
After looking at some example files, I believe I have deciphered at
least the important parts of the headers. Add a document describing
them. Implement reading of the per-track headers and remaining important
bits of the audio header.

There is a method for reading track data, but for now it is just a stub
and isn't called.